+++ zcash_client_backend/CHANGELOG.md
@@ -10,7 +10,10 @@ and this library adheres to Rust's notion of
 ### Added
 - `zcash_client_backend::data_api`:
   - `chain::BlockCache` trait, behind the `sync` feature flag.
-- `zcash_client_backend::fees::ChangeValue::{transparent, shielded}`
+- `zcash_client_backend::fees`:
+  - `ChangeValue::{transparent, shielded}`
+  - `sapling::EmptyBundleView`
+  - `orchard::EmptyBundleView`
 - `zcash_client_backend::scanning`:
   - `testing` module
 - `zcash_client_backend::sync` module, behind the `sync` feature flag.